% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_mi_results}
\alias{create_mi_results}
\title{Create MI Results Object}
\usage{
create_mi_results(results_list)
}
\arguments{
\item{results_list}{List containing rubin_stats and per_imputation components}
}
\value{
An object of class 'mi_results'
}
\description{
Create MI Results Object
}
