% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_predictor_matrix}
\alias{create_predictor_matrix}
\title{Create Predictor Matrix csv}
\usage{
create_predictor_matrix(df, save_file = NULL)
}
\arguments{
\item{df}{The dataframe, including the column names.}

\item{save_file}{Default value is NULL. If a save file is specified, the function will save the created matrix as a .csv file with the name and location specified.}
}
\value{
A predictor matrix with default settings (all predictors used) for the bigMice imputation procedure.
}
\description{
Utility function to create a matrix or a csv File containing the variables names x variables names matrix to be used for specifying the predictor matrix to be used during the bigMice procedure. The csv file can then be opened with a spreadsheet software (e.g. Excel, Google sheets, etc.) to visualize and specify the predictors of each variable more easily and in a friendlier interface.
}
\examples{
#TBD
}
