% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_named_vector}
\alias{get_named_vector}
\title{Get Named vector}
\usage{
get_named_vector(csv_file)
}
\arguments{
\item{csv_file}{Path to the CSV file}
}
\value{
A named vector of the variable types with the variable names as names, ordered alphabetically by variable name.
}
\description{
Utility function to read a CSV file containing Variable Name and Variable Type columns and return a named vector of the variable types with the variable names as names. The vector is ordered alphabetically by variable name.
}
\examples{
#TBD
}
