% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_predictor_matrix}
\alias{import_predictor_matrix}
\title{Import Predictor Matrix csv}
\usage{
import_predictor_matrix(file)
}
\arguments{
\item{file}{The csv file to be opened}
}
\value{
A predictor matrix with specified settings (predictors used) for the bigMice imputation procedure.
}
\description{
Utility function to import back a predictor matrix from a csv file created by the create_predictor_matrix() function.
}
\examples{
#TBD

}
