% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer_mult_logistic.R
\name{impute_with_mult_logistic_regression}
\alias{impute_with_mult_logistic_regression}
\title{Multinomial Logistic Regression Imputation function}
\usage{
impute_with_mult_logistic_regression(sc, sdf, target_col, feature_cols)
}
\arguments{
\item{sc}{A Spark connection}

\item{sdf}{A Spark DataFrame}

\item{target_col}{The column with missing values to impute}

\item{feature_cols}{The columns to use as features in the multinomial logistic regression model. These columns should not have missing values.}
}
\value{
The Spark DataFrame with missing values imputed in the target column
}
\description{
This function imputes missing values in a Spark DataFrame using Multinomial Logistic regression.
}
\examples{
#TBD
}
