% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer_random_forest.R
\name{impute_with_random_forest_regressor}
\alias{impute_with_random_forest_regressor}
\title{Random Forest Regression Imputation function}
\usage{
impute_with_random_forest_regressor(
  sc,
  sdf,
  target_col,
  feature_cols,
  target_col_prev
)
}
\arguments{
\item{sc}{A Spark connection}

\item{sdf}{A Spark DataFrame}

\item{target_col}{The column with missing values to impute}

\item{feature_cols}{The columns to use as features in the Random Forest regression model. These columns should not have missing values.}

\item{target_col_prev}{the target column at the previous iteration. Used to calculate residuals.}
}
\value{
The Spark DataFrame with missing values imputed in the target column
}
\description{
This function imputes missing values in a Spark DataFrame using Random Forest regression.
}
\examples{
#TBD
}
