% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sample_imputation.R
\name{impute_with_random_samples}
\alias{impute_with_random_samples}
\title{Random Sample Imputation function}
\usage{
impute_with_random_samples(sc, sdf, column = NULL)
}
\arguments{
\item{sc}{A Spark connection}

\item{sdf}{A Spark DataFrame}

\item{column}{The column(s) to impute. If NULL, all columns will be imputed}
}
\value{
The Spark DataFrame with missing values imputed
}
\description{
This function imputes missing values in a Spark DataFrame using random samples from the observed values.
}
\examples{
#TBD
}
