% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_template.R
\name{my_function}
\alias{my_function}
\title{Title of the Function}
\usage{
my_function(x)
}
\arguments{
\item{x}{Description of the parameter `x`}
}
\value{
Description of the return value
}
\description{
A brief description of what the function does. To generate the doc, use `devtools::document()`.
}
\examples{
my_function(10)
}
