% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.mi_results}
\alias{print.mi_results}
\title{Print Method for Multiple Imputation Results}
\usage{
\method{print}{mi_results}(x, digits = 4, show_individual = FALSE, ...)
}
\arguments{
\item{x}{A list containing multiple imputation results with Rubin's statistics}

\item{digits}{Number of digits to display for numeric values (default: 4)}

\item{show_individual}{Logical, whether to show individual imputation results (default: FALSE)}

\item{...}{Additional arguments (not currently used)}
}
\description{
Print Method for Multiple Imputation Results
}
