% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.mi_results}
\alias{summary.mi_results}
\title{Summary Method for Multiple Imputation Results}
\usage{
\method{summary}{mi_results}(object, ...)
}
\arguments{
\item{object}{A list containing multiple imputation results}

\item{...}{Additional arguments passed to print method}
}
\description{
Summary Method for Multiple Imputation Results
}
