% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer_linear_regression.R
\name{impute_with_linear_regression}
\alias{impute_with_linear_regression}
\title{Linear Regression Imputation function}
\usage{
impute_with_linear_regression(
  sc,
  sdf,
  target_col,
  feature_cols,
  elastic_net_param = 0,
  target_col_prev
)
}
\arguments{
\item{sc}{A Spark connection}

\item{sdf}{A Spark DataFrame}

\item{target_col}{The column with missing values to impute}

\item{feature_cols}{The columns to use as features in the linear regression model. These columns should not have missing values.}

\item{elastic_net_param}{The elastic net parameter for the linear regression model. Default is 0 (ridge regression)}

\item{target_col_prev}{the target column at the previous iteration. Used to calculate residuals.}
}
\value{
The Spark DataFrame with missing values imputed in the target column
}
\description{
This function imputes missing values in a Spark DataFrame using linear regression.
}
\examples{
#TBD
}
