% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer_mean_mode_median.R
\name{impute_with_MeMoMe}
\alias{impute_with_MeMoMe}
\title{Mean/Mode/Median Imputation function}
\usage{
impute_with_MeMoMe(sc, sdf, column = NULL, impute_mode, printFlags = TRUE)
}
\arguments{
\item{sc}{A Spark connection}

\item{sdf}{A Spark DataFrame}

\item{column}{The column(s) to impute. If NULL, all columns will be imputed}

\item{impute_mode}{Which imputation method to use for each column. Options are "mean", "mode", "median", or "none"}

\item{printFlags}{Wether or not to print the imputation process to the console. Default is TRUE.}
}
\value{
The Spark DataFrame with missing values imputed
}
\description{
This function imputes missing values in a Spark DataFrame using the mean, mode, or median of the observed values.
}
\examples{
#TBD
}
