% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.spark.R
\name{sampler.spark}
\alias{sampler.spark}
\title{MICE sampler function}
\usage{
sampler.spark(
  sc,
  data,
  imp_init,
  fromto,
  var_types,
  predictorMatrix,
  printFlag
)
}
\arguments{
\item{sc}{A Spark connection}

\item{data}{A Spark DataFrame, the original data with missing values}

\item{imp_init}{A Spark DataFrame, the original data with missing values, but with initial imputation (by random sampling or mean/median/mode imputation)}

\item{fromto}{A vector of length 2, the range of iterations to perform (from, to)}

\item{var_types}{A named character vector, the variable types of the columns in the data.}

\item{predictorMatrix}{A matrix, the predictor matrix to use for the imputation. TBD}

\item{printFlag}{A boolean, whether to print debug information.}
}
\value{
The Spark DataFrame with missing values imputed for all variables
}
\description{
This function is the core of the MICE algorithm. It iteratively imputes missing values in a Spark DataFrame using a set of imputation methods based on the variable types.
}
\examples{
#TBD
}
